move split and trim to utils
